添加 查询 产品信息

+++ src/main/java/com/robert/myschool/utils/Pager.java
@@ -17,7 +17,7 @@ public class Pager<T> implements Serializable {
 
   private Integer pageIndex;
   private Integer pageSize;
-  private Integer total;
+  private Long total;
   private Integer pages;
   private List<T> list;
